game-application/backend-FASTAPI-MONGODB/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pymongo
motor